rtmetric - individual named metrics (counters, gauges, strings) rtflow   - sampled transaction record with multiple named fields
examples:
========

{
  "rtmetric": {
    "datasource": "rtm_ds1",
    "metric1": { "type":"counter32", "value": 777 },
    "metric2": { "type":"string", "value":"hello world" },
    "metric3": { "type":"gaugedouble", "value": 123.45 }
  }
}

{
  "rtflow": {
    "datasource":"rtf_ds1",
    "sampling_rate": 4,
    "field1": { "type":"int64", "value": 1e12 },
    "field2": { "type":"mac", "value":"02:03:04:05:06:07" },
    "field3": { "type":"ip6",  "value": "FD03::0102" }
  }
}

rtmetric metric types:
=====================
counter32
counter64
gauge32
gauge64
guagefloat
gaugedouble
string

rtflow field types:
==================
int32
int64
float
double
mac
ip
ip6
string

These rtmetric and rtflow messages are packed into the binary sFlow feed
for efficiency, and to ensure that loss-in-transit can be detected in the
normal way.

field restrictions:
==================
rtmetric metric names and rtflow field names must be <= 64 bytes and
contain only alphanumeric characters or '-' or '_'.  (So metrics with '.'
characters will be rejected).

The "datasource" field can be used to identify separate entities on the
same host that may be sending the same metric.  It has the same length
and character restrictions,  plus it may not start with a digit.

The "string" metrics and fields may contain any characters expressible
in JSON,  but must be <= 255 bytes in length.