In a major reworking, the Linux port is now factored into loadable modules. This makes debugging much easier,  and it also makes it possible to build one .rpm and one .deb for all Linux-variant platforms and feature-sets.
This commit also bumps the version to 2.0.1, and updates all references
to the old sourceforge site to point to http://sflow.net.

Under src/Linux/, the loadable modules are in mod_*.c, and the framework
that loads them and allows them to communicate is in evbus.[ch].  The
evbus framework keeps activity on different threads as separate as possible,
so that in most places the code enjoys the simplicity of being effectively
single-threaded.

The Makefile no longer guesses what features you want to include.  If you
just run "make" it will try to build all the features, but if you run
something like:

  make FEATURES="DOCKER PCAP"

then only mod_docker and mod_pcap will be compiled, along with hsflowd
and the core modules.

The sflowovsd daemon has been subsumed,  and now exists as a module
in hsflowd.   To compile it,  ensure that "OVS" is in the list of features,
and that the etc/hsflowd.conf file includes an "ovs {}" section.

The /etc/hsflowd.conf file is backwards compatible,  but some fields are
deprecated.  For example,  if you have "JSONPort=36343" the new code
will suggest a change to "json { UDPPort=36343 }".