mod_os10 overhaul. Separate pollBus and packetBus activity. Get interface speeds via script. Added counter polling via script.
Modules can get minPollingInterval.  If the config file setting is less then
this is used instead.  Separtately the syncPollingInterval can be set to
cause pollers to cluster together in batches.  This can be used to
reduce the number of datagrams sent by an idle switch,  for example.

"make dist" now adds the release number to the tarball.

Added EVBusExec utility for piping in lines ASCII data from a script
asychronously, tracking stdout and stderr separately, and capturing
the exit status. Not used yet.

Modules can now influence the discovery of interfaces,  controlling
which ethtool structures will be requested.

Sorry for the squashed commit.