mod_pcap use tap device as sampler and allow config option pcap { dev=<name> vport=on } to ask for separate vport counters on every active port.  The vport=on option is implicit if the device being tapped is a bridge. This brings it into line with what happens on Open vSwitch sFlow.
This change was prompted by Docker swarm and the way
that container veth peers do not all show up in the global
namespace (with container-specific MACs in the sampled packets).
This also affected the retrieval of container network counters,
which are now read from /proc/<container-pid>/net/dev instead of
being lifted from the global-namespace ends of the container
veth pairs.