Darwin port: updated to adopt modular appoach.  Although for now the modules are not shared libraries,  but are compiled into hsflowd.
Only mod_json, mod_dnssd and mod_pcap are included.

There are many gaps in the metrics, which could be filled by calling
out to iostat(1), netstat(1) and vm_stat(1) but if it's possible to
do it more directly with calls to sysctl then that is preferred.

This code compiles on MacOS Sierra.  I have not tested on other
OS versions.

There is still no .plist file, and no script for constructing
a .dmg installation file.  That remains a todo item.