Going forward, the build for OS10 is the same as for OPX except that the CPS libraries are found under different paths and the environment vars are set differently in the hsflowd.service file. So this commit removes the old mod_os10.c module.
Config files can still have os10{} sections and they will be
treated the same as opx{} sections.

The Docker build for OS10 is also removed.  It can be replaced
if a docker image is published for OS10 development.

The Makefile still supports "make FEATURES=OS10".  When used this
way the Makefile will build mod_opx with the linking options that
are appropriate for OS10.  It will also substitute a custom
systemd unit file that has the OS10 environment settings.